sb-94 remove custom build scripts

+++ packages/image/lib/format-files.js
@@ -133,14 +133,15 @@ export const formatFiles = async () => {
 
 export const cleanOutDir = async () => {
     const { outputDir } = getDirs(true);
-    const rmQueue = IS_LIMITED_CI
-        ? [
-              fs.rm(path.join(IMAGE_DIR, "node_modules", ".bin"), { recursive: true, force: true }),
-              fs.rm(path.join(IMAGE_DIR, "node_modules", "next"), { recursive: true, force: true }),
-              fs.rm(path.join(IMAGE_DIR, "node_modules", "react"), { recursive: true, force: true }),
-              fs.rm(path.join(IMAGE_DIR, "node_modules", "react-dom"), { recursive: true, force: true }),
-          ]
-        : [fs.rm(path.join(IMAGE_DIR, "node_modules", ".bin"), { recursive: true, force: true })];
+    const rmQueue = [fs.rm(path.join(IMAGE_DIR, "node_modules", ".bin"), { recursive: true, force: true })];
+    if (IS_LIMITED_CI) {
+        rmQueue.push(
+            fs.rm(path.join(IMAGE_DIR, "node_modules", "next"), { recursive: true, force: true }),
+            fs.rm(path.join(IMAGE_DIR, "node_modules", "react"), { recursive: true, force: true }),
+            fs.rm(path.join(IMAGE_DIR, "node_modules", "react-dom"), { recursive: true, force: true }),
+        );
+    }
+
     await Promise.all([
         ...rmQueue,
         fs.rm(path.join(outputDir, "lib"), { recursive: true, force: true }),
@@ -153,10 +154,4 @@ export const cleanOutDir = async () => {
         fs.rm(path.join(outputDir, "tsconfig.json"), { recursive: true, force: true }),
     ]);
     await fs.mkdir(outputDir, { recursive: true });
-};
-
-export const cleanSlides = async () => {
-    const { outputDir } = getDirs(true);
-    await cleanOutDir();
-    await fs.copyFile(path.join(outputDir, "../_slides.ts"), path.join(outputDir, "index.ts"));
 };
